Added unitary_ops header
Added quantum_kernel and quantum expressions for applying unitary
operations and controlled unitary operations to unitary_ops.h.

Added an example to demonstrate and compare results for equivalent
implementations.

Added an example to verify the equivalence between the matrix
representation input to the function fromUnitary1() and the sequence of
rotation gates returned.

Signed-off-by: Kevin Rasch <kevin.rasch@intel.com>